Se implemento la deteccion de huellas, la muestra y la guarda en la DB, y tambien se agrego una clase para insertar cualquier registro a la DB

+++ src/Modelo/Usuario.java
@@ -21,7 +21,7 @@ public class Usuario {
     
     private String nombre;
     private String apellido;
-    private int DPI;
+    private long DPI;
     private String userName;
     private String password;
     private byte huella[];
@@ -45,11 +45,11 @@ public class Usuario {
         this.apellido = apellido;
     }
 
-    public int getDPI() {
+    public long getDPI() {
         return DPI;
     }
 
-    public void setDPI(int DPI) {
+    public void setDPI(long DPI) {
         this.DPI = DPI;
     }
